Migrate to TS Stories
