+++ src/stories/lib/LayoutMX_Auto.svelte
@@ -1,0 +1,3 @@
+<div class="mx-auto ">
+	<slot />
+</div>
